+++ core/src/compression/gzip_benchmark.cpp
@@ -1,20 +1,17 @@
 #include <benchmark/benchmark.h>
 
-#include <random>
 #include <chrono>
+#include <random>
 
-#include <zstd.h>
-#include <compression/zstd.hpp>
 #include <boost/iostreams/filter/gzip.hpp>
 #include <boost/iostreams/filtering_stream.hpp>
 #include <compression/gzip.hpp>
 
-#include <userver/logging/log.hpp>
-
 USERVER_NAMESPACE_BEGIN
 
 std::string GenerateRandomData(std::size_t size) {
-  std::mt19937 random_device(std::chrono::steady_clock::now().time_since_epoch().count());
+  std::mt19937 random_device(
+      std::chrono::steady_clock::now().time_since_epoch().count());
   std::uniform_int_distribution<int> dist(0, 25);
   std::string output = "";
   for (std::size_t ind = 0; ind < size; ++ind) {
@@ -23,30 +20,6 @@ std::string GenerateRandomData(std::size_t size) {
   }
   return output;
 }
-
-static void ZstdDecompress(benchmark::State& state) {
-  for ([[maybe_unused]] auto _ : state) {
-    state.PauseTiming();
-    const std::size_t kSize = state.range(0);
-    auto data = GenerateRandomData(kSize);
-
-    const std::size_t kMaxSize = ZSTD_compressBound(kSize);
-    std::string comp_buf (kMaxSize, '\0');
-
-    const std::size_t kCompSize =
-        ZSTD_compress(comp_buf.data(), kMaxSize, data.data(), kSize, 1);
-
-    if (ZSTD_isError(kCompSize)) {
-      LOG_ERROR() << "Couldn't compress data!";
-      return;
-    }
-
-    state.ResumeTiming();
-    auto decomp_str = compression::zstd::Decompress(
-        std::string_view(comp_buf.data(), kCompSize), kSize);
-  }
-}
-BENCHMARK(ZstdDecompress)->RangeMultiplier(2)->Range(1 << 10, 1 << 15);
 
 static void GzipDecompress(benchmark::State& state) {
   for ([[maybe_unused]] auto _ : state) {
@@ -71,11 +44,9 @@ static void GzipDecompress(benchmark::State& state) {
     }
 
     state.ResumeTiming();
-    auto decomp_str = compression::gzip::Decompress(
-        compressed, 1 << 30);
+    auto decomp_str = compression::gzip::Decompress(compressed, 1 << 30);
   }
 }
 BENCHMARK(GzipDecompress)->RangeMultiplier(2)->Range(1 << 10, 1 << 15);
-
 
 USERVER_NAMESPACE_END
